Re-organize a bit controllers
